requirements.txt
Flask~=3.0.0
python-dotenv~=1.0.0
Flask-Mail
cryptography
pymongo==4.7.3
wtforms~=3.1.0
certifi~=2024.6.2